+++ builder_pkgs/mockito/lib/mockito.dart
@@ -1,2 +1,29 @@
-export 'mockito_no_mirrors.dart';
-export 'src/spy.dart' show spy;
+export 'src/mock.dart'
+    show
+    Mock,
+    named,
+
+    // -- setting behaviour
+    when,
+    any,
+    argThat,
+    captureAny,
+    captureThat,
+    typed,
+    Answering,
+    Expectation,
+    PostExpectation,
+
+    // -- verification
+    verify,
+    verifyInOrder,
+    verifyNever,
+    verifyNoMoreInteractions,
+    verifyZeroInteractions,
+    VerificationResult,
+    Verification,
+
+    // -- misc
+    clearInteractions,
+    reset,
+    logInvocations;