+++ builder_pkgs/mockito/lib/mockito_no_mirrors.dart
@@ -1,29 +1,4 @@
-export 'src/mock.dart'
-    show
-        Mock,
-        named,
+@Deprecated('Import "package:mockito/mockito.dart" instead.')
+library mockito.mockito_no_mirrors;
 
-        // -- setting behaviour
-        when,
-        any,
-        argThat,
-        captureAny,
-        captureThat,
-        typed,
-        Answering,
-        Expectation,
-        PostExpectation,
-
-        // -- verification
-        verify,
-        verifyInOrder,
-        verifyNever,
-        verifyNoMoreInteractions,
-        verifyZeroInteractions,
-        VerificationResult,
-        Verification,
-
-        // -- misc
-        clearInteractions,
-        reset,
-        logInvocations;
+export 'mockito.dart';